backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
pytest==8.0.0
httpx==0.26.0
